+++ src/components/events/event/EventCard.jsx
